+++ tests/conftest.py
@@ -40,14 +40,6 @@ NO_LICENSE_PLATE_IMAGES = ["black.jpg", "t3.jpg"]
 
 CROWD_IMAGES = ["crowd1.jpg", "crowd2.jpg"]
 
-UNDISTORT_BEFORE = ["undistort_before.jpg"]
-CORNER_DATA = ["corners_ok.npz", "image_too_small.npz", "not_in_box.npz"]
-CAMERA_COEFFICIENTS = ["camera_calibration_coeffs.yml"]
-CHECKERBOARD_IMAGES = ["checkerboard1.png", "checkerboard2.png"]
-NO_CHECKERBOARD_IMAGES = ["black.jpg"]
-DETECTED_CORNERS = ["detected_corners.npz"]
-CALIBRATION_DATA = ["calibration_data.npz"]
-
 # Paths
 PKD_DIR = Path(__file__).resolve().parents[1] / "peekingduck"
 TEST_DATA_DIR = PKD_DIR.parent / "tests" / "data"
@@ -240,55 +232,6 @@ def human_video_sequence_with_empty_frames(request):
     # check for based on video content
     yield request.param, sequence
 
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=UNDISTORT_BEFORE)
-def undistort_before(request):
-    yield str(TEST_IMAGES_DIR / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=CORNER_DATA)
-def corner_data(request):
-    yield str(TEST_DATA_DIR / "camera_calibration" / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=CAMERA_COEFFICIENTS)
-def camera_coefficients(request):
-    yield str(TEST_DATA_DIR / "undistort" / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=CHECKERBOARD_IMAGES)
-def checkerboard_images(request):
-    yield str(TEST_IMAGES_DIR / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=NO_CHECKERBOARD_IMAGES)
-def no_checkerboard_images(request):
-    yield str(TEST_IMAGES_DIR / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=DETECTED_CORNERS)
-def detected_corners(request):
-    yield str(TEST_DATA_DIR / "camera_calibration" / request.param)
-    K.clear_session()
-    gc.collect()
-
-
-@pytest.fixture(params=CALIBRATION_DATA)
-def calibration_data(request):
-    yield str(TEST_DATA_DIR / "camera_calibration" / request.param)
     K.clear_session()
     gc.collect()
